Some minor changes in model, Implemented Tour and location services . #5

+++ src/main/java/com/sharashkina_kontora/travel_agency/service/BasicService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 public interface BasicService<T> {
     List<T> findAll();
 
-    Optional<T> findById(Integer id);
+    Optional<T> findById(Long id);
 
     T save(T t);
